+++ Services/Ipadanian_Service.cs
@@ -32,7 +32,7 @@ namespace Padanian_Bank.Services.BankService{
 
 		List<Account> Search(Guid account_id);
 
-		void LogTransaction(Guid account_id, float funds, DateTime timestamp, TransactionType type);
+		void LogTransaction(Guid account_id, double funds, DateTime timestamp, TransactionType type);
 
 		List<Transaction> GetAccountTransactions(Guid account_id);
